Move stories into correct folder structure
