+++ admin/resources/js/stories/H2Components/Button.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Story, Meta } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
-import Button, { ButtonProps } from "../components/H2Components/Button";
+import Button, { ButtonProps } from "../../components/H2Components/Button";
 
 export default {
   component: Button,
